third_party/blink/web_tests/ppapi/permissionclient/plugin-permission-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
PASS: first plugin loaded
